refactor navbar in multiple parts + add navbar item types validation + try to fix remaining merge bugs
